Simplify output in login and project
