+++ pkg/cmd/cli/cmd/loginoptions.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"os"
 	"path/filepath"
 
@@ -11,7 +12,9 @@ import (
 	kclient "github.com/GoogleCloudPlatform/kubernetes/pkg/client"
 	kclientcmd "github.com/GoogleCloudPlatform/kubernetes/pkg/client/clientcmd"
 	clientcmdapi "github.com/GoogleCloudPlatform/kubernetes/pkg/client/clientcmd/api"
+	kclientcmdapi "github.com/GoogleCloudPlatform/kubernetes/pkg/client/clientcmd/api"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/fields"
+	kcmdconfig "github.com/GoogleCloudPlatform/kubernetes/pkg/kubectl/cmd/config"
 	kubecmdconfig "github.com/GoogleCloudPlatform/kubernetes/pkg/kubectl/cmd/config"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/labels"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
@@ -25,6 +28,38 @@ import (
 )
 
 const defaultClusterURL = "https://localhost:8443"
+
+// LoginOptions is a helper for the login and setup process, gathers all information required for a
+// successful login and eventual update of config files.
+// Depending on the Reader present it can be interactive, asking for terminal input in
+// case of any missing information.
+// Notice that some methods mutate this object so it should not be reused. The Config
+// provided as a pointer will also mutate (handle new auth tokens, etc).
+type LoginOptions struct {
+	Server string
+
+	// flags and printing helpers
+	Username string
+	Password string
+	Project  string
+
+	// infra
+	StartingKubeConfig *kclientcmdapi.Config
+	DefaultNamespace   string
+	Config             *kclient.Config
+	Reader             io.Reader
+	Out                io.Writer
+
+	// cert data to be used when authenticating
+	CAFile      string
+	CertFile    string
+	KeyFile     string
+	InsecureTLS bool
+
+	Token string
+
+	PathOptions *kcmdconfig.PathOptions
+}
 
 // Gather all required information in a comprehensive order.
 func (o *LoginOptions) GatherInfo() error {
